iq
h

q
iq
:wq

e commit.

+++ FoodOrderingAPI/Services/PromoCodeService.cs
@@ -1,12 +1,14 @@
 ﻿using AutoMapper;
+using FoodOrderingAPI.Controllers;
 using FoodOrderingAPI.DTO;
 using FoodOrderingAPI.Models;
 using FoodOrderingAPI.Repository;
 using Microsoft.AspNetCore.Identity;
+using NetTopologySuite.Index.HPRtree;
 
 namespace FoodOrderingAPI.Services
 {
-    public class PromoCodeService:IPromoCodeService
+    public class PromoCodeService : IPromoCodeService
     {
         private readonly IWebHostEnvironment _environment;
         private readonly ApplicationDBContext _context;
@@ -30,28 +32,60 @@ namespace FoodOrderingAPI.Services
             return await _repository.AddPromoCodeAsync(restaurantId, promoCode);
         }
 
-        public async Task<PromoCode> UpdatePromoCodeAsync(PromoCode promoCode)
+
+        public async Task<PromoCode?> UpdatePromoCodeAsync(string restaurantId, Guid promoCodeId, PromoCodeDto dto)
         {
-            return await _repository.UpdatePromoCodeAsync(promoCode);
+            var existingCode = await _repository.GetPromoCodeByIdAsync(promoCodeId);
+            if (existingCode == null)
+                return null;
+
+
+            existingCode.IssuedByID = restaurantId;
+            existingCode.IssuedByType = RoleEnum.Restaurant.ToString();
+
+            if (dto.ExpiryDate != null)
+            {
+                existingCode.ExpiryDate = dto.ExpiryDate.Value;
+            }
+
+            if (dto.Code != null)
+            {
+                existingCode.Code = dto.Code;
+            }
+
+            if (dto.DiscountPercentage != null)
+            {
+                existingCode.DiscountPercentage = dto.DiscountPercentage.Value;
+            }
+
+            if (dto.IsFreeDelivery != null)
+            {
+                existingCode.IsFreeDelivery = dto.IsFreeDelivery.Value;
+            }
+
+            if (dto.UsageLimit != null)
+            {
+                existingCode.UsageLimit = dto.UsageLimit.Value;
+            }
+
+            return await _repository.UpdatePromoCodeAsync(existingCode);
+
         }
+
+
+
+
 
         public async Task<bool> DeletePromoCodeAsync(Guid promoCodeId, string restaurantId)
         {
+            var existingCode = await _repository.GetPromoCodeByIdAsync(promoCodeId);
+            if (existingCode == null)
+                return false;
+
             return await _repository.DeletePromoCodeAsync(promoCodeId, restaurantId);
         }
-        public async Task<PromoCode> GetPromoCodesByCodeAsync(string restaurantId, string code)
-        {
-            if (string.IsNullOrWhiteSpace(restaurantId))
-                throw new ArgumentException("Restaurant ID must be provided.", nameof(restaurantId));
 
-            if (string.IsNullOrWhiteSpace(code))
-                throw new ArgumentException("Promocode must be provided.", nameof(code));
 
-            if (!Guid.TryParse(restaurantId, out Guid rid))
-                throw new ArgumentException("Invalid Restaurant ID format.", nameof(restaurantId));
-
-            return await _repository.GetPromoCodesByCodeAsync(restaurantId, code);
-        }
         public async Task<IEnumerable<PromoCode>> GetAllPromoCodesByRestaurantAsync(string restaurantId)
         {
             if (string.IsNullOrWhiteSpace(restaurantId))
@@ -76,66 +110,10 @@ namespace FoodOrderingAPI.Services
 
             return await _repository.SearchPromoCodesByCodeAsync(restaurantId, code);
         }
-        public async Task<PromoCodeApplyDto> ValidatePromoCode(string restaurantId, string code,decimal subtotal,decimal delivaryPrice)
-        {
-            PromoCode pcode;
-            try
-            {
-                pcode = await GetPromoCodesByCodeAsync(restaurantId, code);
-            }
-            catch(Exception ex)
-            {
-                return new PromoCodeApplyDto
-                {
-                    IsVaild = false,
-                    message = ex.Message
-                };
-            }
-            if (pcode == null)
-                return new PromoCodeApplyDto
-                {
-                    IsVaild = false,
-                    message = "this promocode Not Found"
-                };
-            if(DateTime.Now> pcode.ExpiryDate)
-                return new PromoCodeApplyDto
-                {
-                    IsVaild = false,
-                    message = "promocode expired cannot be applied"
-                };
-            if(await _repository.getPromoCodeUses(code)>= pcode.UsageLimit && pcode.UsageLimit>0)
-                return new PromoCodeApplyDto
-                {
-                    IsVaild = false,
-                    message = "This promo code has reached its maximum usage limit. "
-                };
-            return new PromoCodeApplyDto
-            {
-                IsVaild = true,
-                PromoCodeID = pcode.PromoCodeID,
-                SubTotal = subtotal,
-                DiscountAmount = ((decimal)pcode.DiscountPercentage * subtotal) / 100,
-                DelivaryPrice = pcode.IsFreeDelivery ? delivaryPrice : 0,
-                message="PromoCode applied successfully"
-            };
-        }
-        public async Task<bool> ApplyPromoCode(Order order)
-        {
-            PromoCode pcode = await _repository.GetPromoCodesById((Guid)order.PromoCodeID);
-        if (pcode == null)
-            {
-                return false;
-            }
-            PromoCodeApplyDto result = await ValidatePromoCode(order.RestaurantID, pcode.Code, order.SubTotal, order.DelivaryPrice);
-            if (result.IsVaild)
-            {
-                order.SubTotal = result.SubTotal;
-                order.DiscountAmount = result.DiscountAmount;
-                order.DelivaryPrice = result.DelivaryPrice;
-                return true;
-            }
-            return false;
-        }
 
+        public async Task<PromoCode?> GetPromoCodeByIdAsync(Guid promoCodeId)
+        {
+            return await _repository.GetPromoCodeByIdAsync(promoCodeId);
+        }
     }
 }
